# Enhancement: Speed-up task reordering for large projects