# Pull Request
- addresses: https://github.com/coleam00/Archon/issues/456

## Summary
Fixes N+1 database query performance bottleneck in task creation that was causing 18+ second response times, reducing API performance by 97.8%.

## Changes Made
- Replaced N+1 UPDATE query pattern in task reordering logic with single bulk UPDATE operation
- Added PostgreSQL function `increment_task_order()` for efficient bulk task reordering
- Created migration script `001_task_reorder_performance_fix.sql` for existing installations
- Updated `complete_setup.sql` to include the optimization for new installations

## Type of Change
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] Performance improvement

## Affected Services
- [x] Server (FastAPI backend)
- [x] Database (migrations/schema)

## Testing
- [x] All existing tests pass
- [x] Manually tested affected user flows

### Test Evidence
```bash
# Task-related unit tests all passing
cd python && uv run pytest tests/ -k "task" -v
# Result: 32 passed, 235 deselected

# Performance testing results
# Before: 18.762 seconds per task creation
# After:  0.411 seconds per task creation
time curl -X POST http://localhost:8181/api/tasks \
  -H "Content-Type: application/json" \
  -d '{"project_id": "...", "title": "Test", "task_order": 1}'
# Result: 0.411s total (45x improvement)
```

## Checklist
- [x] My code follows the service architecture patterns
- [x] All new and existing tests pass locally
- [x] My changes generate no new warnings
- [x] I have verified no regressions in existing features

## Additional Notes
### Performance Metrics
- **Before**: 18.762 seconds per task creation (N+1 queries)
- **After**: 0.411 seconds per task creation (single UPDATE)  
- **Improvement**: 97.8% reduction in response time (45x faster)

### Root Cause
The task reordering logic was executing individual UPDATE queries for each existing task that needed to be reordered, creating an N+1 query pattern. With many tasks in a project, this caused severe performance degradation.

### Solution
Implemented a PostgreSQL function that performs the reordering in a single UPDATE statement, eliminating the expensive loop of individual database calls.

### Migration Instructions
For existing installations, run the migration in Supabase SQL Editor:
```sql
-- Run migration/001_task_reorder_performance_fix.sql
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- Bug Fixes
  - Resolved slowdowns when creating tasks at specific positions. Task creation and reordering now complete significantly faster, especially in larger projects.

- Refactor
  - Optimized task reordering behind the scenes to a single operation with retry/backoff for improved reliability under load. No changes to UI or workflow.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->