@ericfisherdev 2025-08-24
closing because I may have found an easier solution